Extract lsp-proto crate
