+++ crates/lsp-engine/src/events.rs
@@ -1,6 +1,4 @@
-use std::path::PathBuf;
-
-use ls_types::{Diagnostic, NumberOrString};
+use rg_lsp_proto::EngineEvent;
 use tokio::sync::mpsc;
 
 pub type EngineEventReceiver = mpsc::UnboundedReceiver<EngineEvent>;
@@ -25,37 +23,4 @@ impl EngineEventSink {
             tracing::debug!("dropped engine event because LSP receiver is closed");
         }
     }
-}
-
-/// Engine-originated side effect that the LSP orchestrator should publish to the client.
-#[derive(Debug)]
-pub enum EngineEvent {
-    PublishDiagnostics {
-        path: PathBuf,
-        diagnostics: Vec<Diagnostic>,
-        version: Option<i32>,
-    },
-    BeginWorkDoneProgress {
-        token: NumberOrString,
-        title: String,
-        message: Option<String>,
-    },
-    EndWorkDoneProgress {
-        token: NumberOrString,
-        message: Option<String>,
-    },
-    InlayHintRefresh,
-    LogMessage {
-        level: EngineLogLevel,
-        message: String,
-    },
-}
-
-/// Client-facing log severity requested by the engine.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum EngineLogLevel {
-    Error,
-    Warning,
-    Info,
-    Log,
 }
