+++ crates/lsp-engine/src/lib.rs
@@ -1,3 +1,9 @@
+//! Analysis-engine implementation for Rust Glancer LSP sessions.
+//!
+//! This crate owns workspace analysis, document freshness tracking, diagnostics execution, and
+//! memory reporting for an engine instance. Public APIs expose engine construction and event
+//! delivery primitives; shared request and notification contracts live in `rg_lsp_proto`.
+
 mod check;
 mod documents;
 mod engine;
@@ -7,11 +13,7 @@ mod project_stats;
 mod proto;
 
 pub use self::{
-    check::CheckConfig,
-    engine::{
-        EngineNotifyFuture, EngineResultFuture, EngineService, EngineServiceHandle,
-        InProcessEngineService,
-    },
-    events::{EngineEvent, EngineEventReceiver, EngineEventSink, EngineLogLevel},
+    engine::InProcessEngineService,
+    events::{EngineEventReceiver, EngineEventSink},
     memory::{AllocatorPurgeResult, AllocatorStats, MemoryControl},
 };
