+++ crates/lsp-proto/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rg_lsp_engine"
+name = "rg_lsp_proto"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -12,14 +12,6 @@ description.workspace = true
 
 [dependencies]
 anyhow.workspace = true
-cargo_metadata.workspace = true
-derive_more = { workspace = true, features = ["debug", "display"] }
 ls_types.workspace = true
-rg_analysis.workspace = true
-rg_def_map.workspace = true
-rg_lsp_proto.workspace = true
-rg_parse.workspace = true
 rg_project.workspace = true
 rg_workspace.workspace = true
-tokio = { workspace = true, features = ["process", "sync"] }
-tracing.workspace = true
